% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvironImport.R
\name{EnvironImport}
\alias{EnvironImport}
\title{Import a weather station file created by an Environdata weather station}
\usage{
EnvironImport(file)
}
\arguments{
\item{file}{Filename of the csv file.}
}
\value{
List with three name elements: Data with the actual data as data.frame, Parameter names and Units for each parameter. Grand totals, means etc provided in the Environdata file are currently discarded.
}
\description{
Import a weather station file created by an Environdata weather station
}
\examples{
\dontrun{
weather <- EnvironImport(myfile.csv)
head(weather$Data)
weather$Parameters
weather$Units
}
}

